This time, 30,000 pieces of data were processed
There are three databases, IC, GPS,bus_station, and the structure of the table are shown in the above picture. The steps that need to be done are also shown in the picture. A brief description of the process of using mysql and python to filter data is given,First, calculate the difference of time field through IC table, and then compare it with the given value a to screen out all the data whose difference of time is less than a. This step adopts dataframe in python, which is more convenient.Second, by mysql to connect IC table and GPS table, more complex requirements for calculating the minimum difference of same time fields in the database, using the python powerful shift function of pandas, again through the grouping and aggregation in the mysql can query to meet the requirements of the data, finally through bus_station connected to the primary key of the GPS to filter the final data exist to the new table